optimize storage

+++ contracts/Pool.sol
@@ -51,9 +51,9 @@ abstract contract Pool is IPool, Storage, Events, Constants {
         uint R = IERC20(TOKEN_R).balanceOf(address(this));
         require(params.a <= R >> 1 && params.b <= R >> 1, "IP");
 
-        s_i = INIT_TIME;
-        s_a = params.a;
-        s_b = params.b;
+        s_i = uint32(INIT_TIME);
+        s_a = uint224(params.a);
+        s_b = uint224(params.b);
 
         uint idA = _packID(address(this), SIDE_A);
         uint idB = _packID(address(this), SIDE_B);
