+++ contracts/logics/Storage.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.0;
 
 contract Storage {
-    uint internal s_i; // TODO: storag optimization
-    uint internal s_a;
-    uint internal s_b;
+    uint32  internal s_i;
+    uint224 internal s_a;
+    uint224 internal s_b;
 }
